# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDM Invasion-Risk API",
    "description": "Species-keyed MLOps service for MaxEnt species distribution models: incremental GBIF/iNaturalist ingestion, conditional retraining, model lineage, map-layer descriptors (titiler / pg_tileserv) and automated bulletins.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/species/{taxon_key}/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "description": "All model versions, newest first, with full reproducibility metadata.",
        "operationId": "listModelVersions",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelVersionOut"
                  },
                  "title": "Response Listmodelversions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/models/{model_version}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Model",
        "operationId": "getModelVersion",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          },
          {
            "name": "model_version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FoldMetricsOut": {
        "properties": {
          "fold": {
            "type": "integer",
            "title": "Fold"
          },
          "n_train_presence": {
            "type": "integer",
            "title": "N Train Presence"
          },
          "n_test_presence": {
            "type": "integer",
            "title": "N Test Presence"
          },
          "n_test_background": {
            "type": "integer",
            "title": "N Test Background"
          },
          "auc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auc"
          },
          "tss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tss"
          },
          "cbi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cbi"
          }
        },
        "type": "object",
        "required": [
          "fold",
          "n_train_presence",
          "n_test_presence",
          "n_test_background",
          "auc",
          "tss",
          "cbi"
        ],
        "title": "FoldMetricsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelVersionDetail": {
        "properties": {
          "model_version": {
            "type": "integer",
            "title": "Model Version"
          },
          "model_type": {
            "type": "string",
            "enum": [
              "A",
              "B"
            ],
            "title": "Model Type"
          },
          "confidence_label": {
            "type": "string",
            "enum": [
              "standard",
              "lower"
            ],
            "title": "Confidence Label"
          },
          "trained_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Trained Ts"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "auc_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auc Mean"
          },
          "auc_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auc Std"
          },
          "cbi_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cbi Mean"
          },
          "cbi_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cbi Std"
          },
          "tss_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tss Mean"
          },
          "tss_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tss Std"
          },
          "n_presence": {
            "type": "integer",
            "title": "N Presence"
          },
          "n_background": {
            "type": "integer",
            "title": "N Background"
          },
          "n_cv_folds": {
            "type": "integer",
            "title": "N Cv Folds"
          },
          "feature_classes": {
            "type": "string",
            "title": "Feature Classes"
          },
          "beta_multiplier": {
            "type": "number",
            "title": "Beta Multiplier"
          },
          "severity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity Score"
          },
          "mlflow_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Run Id"
          },
          "reproducibility": {
            "$ref": "#/components/schemas/ReproducibilityInfo"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "threshold_rule": {
            "type": "string",
            "title": "Threshold Rule"
          },
          "cv_method": {
            "type": "string",
            "title": "Cv Method"
          },
          "predictors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Predictors"
          },
          "variable_importance": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Variable Importance"
          },
          "folds": {
            "items": {
              "$ref": "#/components/schemas/FoldMetricsOut"
            },
            "type": "array",
            "title": "Folds"
          },
          "tuning": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tuning"
          },
          "projection": {
            "$ref": "#/components/schemas/ProjectionSummaryOut"
          },
          "transferability_caveat": {
            "type": "string",
            "title": "Transferability Caveat"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeverityResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "n_records_by_effective_label": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "N Records By Effective Label"
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "model_type",
          "confidence_label",
          "trained_ts",
          "trigger",
          "n_presence",
          "n_background",
          "n_cv_folds",
          "feature_classes",
          "beta_multiplier",
          "reproducibility",
          "threshold",
          "threshold_rule",
          "cv_method",
          "predictors",
          "variable_importance",
          "folds",
          "tuning",
          "projection",
          "transferability_caveat",
          "n_records_by_effective_label"
        ],
        "title": "ModelVersionDetail"
      },
      "ModelVersionOut": {
        "properties": {
          "model_version": {
            "type": "integer",
            "title": "Model Version"
          },
          "model_type": {
            "type": "string",
            "enum": [
              "A",
              "B"
            ],
            "title": "Model Type"
          },
          "confidence_label": {
            "type": "string",
            "enum": [
              "standard",
              "lower"
            ],
            "title": "Confidence Label"
          },
          "trained_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Trained Ts"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "auc_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auc Mean"
          },
          "auc_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auc Std"
          },
          "cbi_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cbi Mean"
          },
          "cbi_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cbi Std"
          },
          "tss_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tss Mean"
          },
          "tss_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tss Std"
          },
          "n_presence": {
            "type": "integer",
            "title": "N Presence"
          },
          "n_background": {
            "type": "integer",
            "title": "N Background"
          },
          "n_cv_folds": {
            "type": "integer",
            "title": "N Cv Folds"
          },
          "feature_classes": {
            "type": "string",
            "title": "Feature Classes"
          },
          "beta_multiplier": {
            "type": "number",
            "title": "Beta Multiplier"
          },
          "severity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity Score"
          },
          "mlflow_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Run Id"
          },
          "reproducibility": {
            "$ref": "#/components/schemas/ReproducibilityInfo"
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "model_type",
          "confidence_label",
          "trained_ts",
          "trigger",
          "n_presence",
          "n_background",
          "n_cv_folds",
          "feature_classes",
          "beta_multiplier",
          "reproducibility"
        ],
        "title": "ModelVersionOut"
      },
...
      "ProjectionSummaryOut": {
        "properties": {
          "land_area_km2": {
            "type": "number",
            "title": "Land Area Km2"
          },
          "nonnative_land_area_km2": {
            "type": "number",
            "title": "Nonnative Land Area Km2"
          },
          "native_suitable_area_km2": {
            "type": "number",
            "title": "Native Suitable Area Km2"
          },
          "established_outside_area_km2": {
            "type": "number",
            "title": "Established Outside Area Km2"
          },
          "candidate_area_km2": {
            "type": "number",
            "title": "Candidate Area Km2"
          },
          "candidate_mean_suitability": {
            "type": "number",
            "title": "Candidate Mean Suitability"
          },
          "candidate_fraction_of_nonnative_land": {
            "type": "number",
            "title": "Candidate Fraction Of Nonnative Land"
          },
          "candidate_mess_ok_fraction": {
            "type": "number",
            "title": "Candidate Mess Ok Fraction"
          },
          "candidate_envelope_fraction": {
            "type": "number",
            "title": "Candidate Envelope Fraction"
          },
          "extrapolation_area_km2": {
            "type": "number",
            "title": "Extrapolation Area Km2"
          },
          "top_candidate_regions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top Candidate Regions"
          },
          "top_limiting_variables": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top Limiting Variables"
          }
        },
        "type": "object",
        "required": [
          "land_area_km2",
          "nonnative_land_area_km2",
          "native_suitable_area_km2",
          "established_outside_area_km2",
          "candidate_area_km2",
          "candidate_mean_suitability",
          "candidate_fraction_of_nonnative_land",
          "candidate_mess_ok_fraction",
          "candidate_envelope_fraction",
          "extrapolation_area_km2",
          "top_candidate_regions",
          "top_limiting_variables"
        ],
        "title": "ProjectionSummaryOut"
      },
...
      "ReproducibilityInfo": {
        "properties": {
          "bioclim_version_used": {
            "type": "string",
            "title": "Bioclim Version Used"
          },
          "bioclim_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bioclim Source"
          },
          "bioclim_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bioclim Sha256"
          },
          "bioclim_dvc_md5": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bioclim Dvc Md5",
            "description": "DVC hash of the bioclim version directory (data/bioclim/<v>.dvc)"
          },
          "gbif_download_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gbif Download Key"
          },
          "gbif_download_doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gbif Download Doi"
          },
          "gbif_citation": {
            "type": "string",
            "title": "Gbif Citation"
          },
          "inat_citation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inat Citation"
          },
          "package_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Package Versions"
          },
          "random_seeds": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Random Seeds"
          },
          "training_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Training Config"
          },
          "native_range_source": {
            "type": "string",
            "title": "Native Range Source"
          },
          "native_range_confirmed_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Native Range Confirmed Ts"
          },
          "training_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrainingSnapshotInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact MaxEnt input (presences + background with predictors); absent for versions trained before snapshots were introduced"
          },
          "occurrence_manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrainingSnapshotInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "All valid records considered, with ingested and effective labels"
          }
        },
        "type": "object",
        "required": [
          "bioclim_version_used",
          "gbif_citation",
          "package_versions",
          "random_seeds",
          "training_config",
          "native_range_source"
        ],
        "title": "ReproducibilityInfo"
      },
...
      "SeverityComponents": {
        "properties": {
          "suitability": {
            "type": "number",
            "title": "Suitability"
          },
          "climate_analogy": {
            "type": "number",
            "title": "Climate Analogy"
          },
          "spread_rate": {
            "type": "number",
            "title": "Spread Rate"
          },
          "ecological_impact_prior": {
            "type": "number",
            "title": "Ecological Impact Prior"
          }
        },
        "type": "object",
        "required": [
          "suitability",
          "climate_analogy",
          "spread_rate",
          "ecological_impact_prior"
        ],
        "title": "SeverityComponents"
      },
      "SeverityConfidence": {
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "low",
              "moderate",
              "high"
            ],
            "title": "Level"
          },
          "mess_ok_fraction": {
            "type": "number",
            "title": "Mess Ok Fraction"
          },
          "model_type": {
            "type": "string",
            "enum": [
              "A",
              "B"
            ],
            "title": "Model Type"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "level",
          "mess_ok_fraction",
          "model_type",
          "note"
        ],
        "title": "SeverityConfidence"
      },
      "SeverityConfigModel": {
        "properties": {
          "weights": {
            "$ref": "#/components/schemas/SeverityWeights"
          },
          "impact_prior": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Impact Prior",
            "default": 0.5
          },
          "impact_prior_source": {
            "type": "string",
            "title": "Impact Prior Source",
            "default": "default (unknown)"
          },
          "suitability_saturation": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Suitability Saturation",
            "default": 0.1
          },
          "spread_half_saturation": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Spread Half Saturation",
            "default": 5.0
          },
          "spread_window_years": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 2.0,
            "title": "Spread Window Years",
            "default": 10
          }
        },
        "type": "object",
        "title": "SeverityConfigModel"
      },
      "SeverityResult": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "score_0_100": {
            "type": "number",
            "title": "Score 0 100"
          },
          "category": {
            "type": "string",
            "enum": [
              "low",
              "moderate",
              "high",
              "very high"
            ],
            "title": "Category"
          },
          "components": {
            "$ref": "#/components/schemas/SeverityComponents"
          },
          "weights": {
            "$ref": "#/components/schemas/SeverityComponents"
          },
          "contributions": {
            "$ref": "#/components/schemas/SeverityComponents"
          },
          "descriptions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Descriptions"
          },
          "confidence": {
            "$ref": "#/components/schemas/SeverityConfidence"
          },
          "formula": {
            "type": "string",
            "title": "Formula"
          },
          "config": {
            "$ref": "#/components/schemas/SeverityConfigModel"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs"
          }
        },
        "type": "object",
        "required": [
          "score",
          "score_0_100",
          "category",
          "components",
          "weights",
          "contributions",
          "descriptions",
          "confidence",
          "formula",
          "config",
          "inputs"
        ],
        "title": "SeverityResult"
      },
      "SeverityWeights": {
        "properties": {
          "suitability": {
            "type": "number",
            "minimum": 0.0,
            "title": "Suitability",
            "default": 0.35
          },
          "climate_analogy": {
            "type": "number",
            "minimum": 0.0,
            "title": "Climate Analogy",
            "default": 0.25
          },
          "spread_rate": {
            "type": "number",
            "minimum": 0.0,
            "title": "Spread Rate",
            "default": 0.2
          },
          "ecological_impact_prior": {
            "type": "number",
            "minimum": 0.0,
            "title": "Ecological Impact Prior",
            "default": 0.2
          }
        },
        "type": "object",
        "title": "SeverityWeights"
      },
...
      "TrainingSnapshotInfo": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Artifact path, relative to the artifact root"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "n_rows": {
            "type": "integer",
            "title": "N Rows"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          }
        },
        "type": "object",
        "required": [
          "path",
          "sha256",
          "n_rows",
          "bytes"
        ],
        "title": "TrainingSnapshotInfo",
        "description": "A per-version, content-addressed data file (see training_snapshot.py)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}